+++ packages/javascript/src/index.ts
@@ -4,4 +4,4 @@ export { readCache, readCacheSync } from './read-cache.js';
 export { setLogLevel, setLogger } from './utils/logger.js';
 export { CachettaError } from './errors.js';
 export { hash } from './hash.js';
-export type { CacheConfig, CacheInfo, PathFn, CachableFunction, CachableFunctionSync, Logger, LogLevel } from './types.js';
+export type { CacheConfig, CacheInfo, ClearOptions, PathFn, CachableFunction, CachableFunctionSync, Logger, LogLevel } from './types.js';